requirements.txt
# pip install -r requirements.txt
# pip install mysqlclient pymysql.txt
# pip install djangorestframework

django == 4.0.1
pillow
django-admin-interface
django-crispy-forms
mysqlclient
pymysql
djangorestframework
